+++ internal/enum/validate.go
@@ -2,7 +2,7 @@ package enum
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
@@ -11,6 +11,6 @@ func Validate[T valueser[T]]() schema.SchemaValidateDiagFunc {
 	return validation.ToDiagFunc(validation.StringInSlice(Values[T](), false))
 }
 
-func FrameworkValidate[T valueser[T]]() tfsdk.AttributeValidator {
+func FrameworkValidate[T valueser[T]]() validator.String {
 	return stringvalidator.OneOf(Values[T]()...)
 }